Extracted router into new router.js. imported and utilized in vue instance in main.js
