+++ utilizations.md
@@ -5,3 +5,7 @@ normalize.css
 ## resources
 
 [implementing css reset within vue](https://www.educative.io/courses/learn-vuejs-from-scratch/YQW33n6DD0K)
+
+[vuerouter](https://router.vuejs.org/guide/#html)
+	- [runtimecompiler error](https://github.com/vuejs/vue-cli/issues/2754)
+		- [fix](https://cli.vuejs.org/config/#lintonsave)
